# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API documentation for Laravel project",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Laravel API Server"
        }
    ],
    "paths": {
        "/api/account/request-deletion": {
            "post": {
                "tags": [
                    "Account Deletion"
                ],
                "summary": "Request account deletion (starts 7-day countdown)",
                "description": "User requests to delete account. It will be deleted permanently after 7 days.",
                "operationId": "3c9901039ffae7969803393d376f8826",
                "responses": {
                    "200": {
                        "description": "Account deletion countdown started",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Account delete request successful. Account will be permanently deleted after 7 days."
                                        },
                                        "data": {
                                            "properties": {
                                                "acc_status": {
                                                    "type": "string",
                                                    "example": "pending_deletion"
                                                },
                                                "delete_scheduled_at": {
                                                    "type": "string",
                                                    "example": "2025-01-15T12:00:00Z"
                                                },
                                                "profile": {
                                                    "$ref": "#/components/schemas/UserProfile"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Deletion already requested or account already deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Account deletion is already pending."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/account/deletion-status": {
            "get": {
                "tags": [
                    "Account Deletion"
                ],
                "summary": "Get current account deletion status",
                "description": "Returns deletion countdown, scheduled deletion date, and profile info.",
                "operationId": "a95ec780dd30f6f029b107784dd5e335",
                "responses": {
                    "200": {
                        "description": "Successfully fetched account status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully get the account status"
                                        },
                                        "data": {
                                            "properties": {
                                                "acc_status": {
                                                    "type": "string",
                                                    "example": "pending_deletion"
                                                },
                                                "delete_requested_at": {
                                                    "type": "string",
                                                    "example": "2025-01-10T12:00:00Z"
                                                },
                                                "delete_scheduled_at": {
                                                    "type": "string",
                                                    "example": "2025-01-17T12:00:00Z"
                                                },
                                                "days_remaining": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "profile": {
                                                    "$ref": "#/components/schemas/UserProfile"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/account/cancel-deletion": {
            "post": {
                "tags": [
                    "Account Deletion"
                ],
                "summary": "Cancel pending deletion request",
                "description": "Cancels the 7-day scheduled deletion process.",
                "operationId": "f17d2cce08d2d8419e4a83fc37f65a00",
                "responses": {
                    "200": {
                        "description": "Successfully canceled account deletion"
                    },
                    "400": {
                        "description": "No pending deletion request found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserProfile": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "example": "USR123456"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "Sharif Uddin"
                    },
                    "email": {
                        "type": "string",
                        "example": "example@gmail.com"
                    },
                    "phone": {
                        "properties": {
                            "country_code": {
                                "type": "string",
                                "example": "+880"
                            },
                            "number": {
                                "type": "string",
                                "example": "1850000000"
                            }
                        },
                        "type": "object"
                    },
                    "profile_img": {
                        "type": "string",
                        "example": "https://domain.com/uploads/profile.png"
                    },
                    "auth_type": {
                        "type": "string",
                        "example": "email"
                    },
                    "app_version": {
                        "type": "string",
                        "example": "1.0.5"
                    },
                    "acc_status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2025-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}